navigation controls initial test
