--- NYCApp/NYCApp/MapView.swift
@@ -1,30 +1,0 @@
-//
-//  MapView.swift
-//  NYCApp
-//
-//  Created by Steven Li on 2/11/20.
-//  Copyright © 2020 Andy Zheng. All rights reserved.
-//
-
-import UIKit
-
-class MapView: UIViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-
-        // Do any additional setup after loading the view.
-    }
-    
-
-    /*
-    // MARK: - Navigation
-
-    // In a storyboard-based application, you will often want to do a little preparation before navigation
-    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
-        // Get the new view controller using segue.destination.
-        // Pass the selected object to the new view controller.
-    }
-    */
-
-}
