+++ NYCApp/NYCApp/MapViewController.swift
@@ -1,5 +1,5 @@
 //
-//  MapViewController.swift
+//  MapView.swift
 //  NYCApp
 //
 //  Created by Steven Li on 2/11/20.
@@ -7,24 +7,74 @@
 //
 
 import UIKit
+import MapKit
+import CoreLocation
 
 class MapViewController: UIViewController {
-
+    @IBOutlet weak var mapView: MKMapView!
+        
+    let locationManager = CLLocationManager()
+    let zoomInMeters: Double = 10000
+        
     override func viewDidLoad() {
         super.viewDidLoad()
-
-        // Do any additional setup after loading the view.
+        checkLocationServices()
+            // Do any additional setup after loading the view.
+    }
+        
+    func setUpLocationManager() {
+        locationManager.delegate = self
+        locationManager.desiredAccuracy = kCLLocationAccuracyBest
     }
     
-
-    /*
-    // MARK: - Navigation
-
-    // In a storyboard-based application, you will often want to do a little preparation before navigation
-    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
-        // Get the new view controller using segue.destination.
-        // Pass the selected object to the new view controller.
+    func centerViewOnUserLocation() {
+        if let location = locationManager.location?.coordinate{
+            let region = MKCoordinateRegion.init(center: location, latitudinalMeters: zoomInMeters, longitudinalMeters: zoomInMeters)
+            mapView.setRegion(region, animated: true)
+        }
     }
-    */
+    
+    func checkLocationServices() {
+        if CLLocationManager.locationServicesEnabled(){
+            setUpLocationManager()
+            checkLocationAuthorization()
+        }
+        else{
+            
+        }
+    }
+    
+    func checkLocationAuthorization(){
+        switch CLLocationManager.authorizationStatus() {
+        case .authorizedWhenInUse:
+            mapView.showsUserLocation = true
+            centerViewOnUserLocation()
+            locationManager.startUpdatingLocation()
+        case .denied:
+            break
+        case .notDetermined:
+            locationManager.requestWhenInUseAuthorization()
+        case .restricted:
+            break
+        case .authorizedAlways:
+            break
+        @unknown default:
+            break
+        }
+    }
+}
 
+extension MapViewController: CLLocationManagerDelegate {
+    
+    func locationManager(_ manager: CLLocationManager, didUpdateLocations locations: [CLLocation]) {
+        guard let location = locations.last else { return }
+        let center = CLLocationCoordinate2D(latitude: location.coordinate.latitude, longitude: location.coordinate.longitude)
+        print(center)
+        let region = MKCoordinateRegion.init(center: center, latitudinalMeters: zoomInMeters, longitudinalMeters: zoomInMeters)
+        mapView.setRegion(region, animated: true)
+    }
+    
+    func locationManager(_ manager: CLLocationManager, didChangeAuthorization status: CLAuthorizationStatus) {
+        checkLocationAuthorization()
+    }
 }
