+++ NYCApp/NYCApp/ViewController.swift
@@ -7,74 +7,35 @@
 //
 
 import UIKit
-import MapKit
-import CoreLocation
 
 class ViewController: UIViewController {
 
-    @IBOutlet weak var mapView: MKMapView!
+    @IBOutlet weak var trainAndBusViewControl: UISegmentedControl!
+    @IBOutlet weak var testlabel: UILabel!
     
-    let locationManager = CLLocationManager()
-    let zoomInMeters: Double = 10000
+    @IBOutlet var trainMainView: UIView!
     
+    let storyBoard: UIStoryboard = UIStoryboard(name: "Main", bundle: nil)
+
     override func viewDidLoad() {
         super.viewDidLoad()
-        checkLocationServices()
+        toggleMode(trainAndBusViewControl)
         // Do any additional setup after loading the view.
     }
-    
-    func setUpLocationManager() {
-        locationManager.delegate = self
-        locationManager.desiredAccuracy = kCLLocationAccuracyBest
-    }
-    
-    func centerViewOnUserLocation() {
-        if let location = locationManager.location?.coordinate{
-            let region = MKCoordinateRegion.init(center: location, latitudinalMeters: zoomInMeters, longitudinalMeters: zoomInMeters)
-            mapView.setRegion(region, animated: true)
-        }
-    }
-    
-    func checkLocationServices() {
-        if CLLocationManager.locationServicesEnabled(){
-            setUpLocationManager()
-            checkLocationAuthorization()
-        }
-        else{
-            
-        }
-    }
-    
-    func checkLocationAuthorization(){
-        switch CLLocationManager.authorizationStatus() {
-        case .authorizedWhenInUse:
-            mapView.showsUserLocation = true
-            centerViewOnUserLocation()
-            locationManager.startUpdatingLocation()
-        case .denied:
+    @IBAction func toggleMode(_ sender: UISegmentedControl){
+        switch trainAndBusViewControl.selectedSegmentIndex{
+        case 0:
+            self.navigationItem.title = "Trains"
             break
-        case .notDetermined:
-            locationManager.requestWhenInUseAuthorization()
-        case .restricted:
+        case 1:
+            self.navigationItem.title = "Buses"
+            let MapViewController = self.storyboard!.instantiateViewController(withIdentifier: "MapView") as! MapViewController
+            self.navigationController!.pushViewController(MapViewController, animated: true)
             break
-        case .authorizedAlways:
-            break
-        @unknown default:
+        default:
             break
         }
     }
+
 }
 
-extension ViewController: CLLocationManagerDelegate {
-    
-    func locationManager(_ manager: CLLocationManager, didUpdateLocations locations: [CLLocation]) {
-        guard let location = locations.last else { return }
-        let center = CLLocationCoordinate2D(latitude: location.coordinate.latitude, longitude: location.coordinate.longitude)
-        let region = MKCoordinateRegion.init(center: center, latitudinalMeters: zoomInMeters, longitudinalMeters: zoomInMeters)
-        mapView.setRegion(region, animated: true)
-    }
-    
-    func locationManager(_ manager: CLLocationManager, didChangeAuthorization status: CLAuthorizationStatus) {
-        checkLocationAuthorization()
-    }
-}
